Little refactory to remove method from Klein class
